pyproject.toml
[tool.poetry]
name = "gestao-api"
version = "0.1.0"
description = "Api para registro de tarefas e horas trabalhadas"
authors = ["Diniz <matheus.oliveira@hepta.com.br>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.2"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
